Added Top Users by Location
Used the existing topLocationService, to get top 25 users for the given location,

still will need to use some google api to check what the location is, i.e. city, state, country.

If that gets implemented we will need to add a parameter for topLocationService to accept different localities

as per usual 

beta-gitnear.azurewebsites.net

shows the latest updates without marios fancy front end. :)

Top Location will not work if the user is accessing the site on a network that disable anonymizers